refactor resource listers
This changes the interface for registering resources. Previously it was
necessary to update the `GetListers` function in `resources/listers.go`.
Now every resource registers itself inside of an `init` function which
is defined in `resources/interface.go` (eg `register("IAMUser",
ListIAMUsers)`). Since it has to provide a name for registering it isn't
required anymore to get the resource type via reflections.

Also we are getting rid of the `FooNuke` types in `resources/types.go`,
which just carried the session and don't really give a benefit. Now, the
`ResourceLister` interface changed so it takes a session as a command.
Therefore the intermediate types aren't required anymore.