feat: adding support for nuking aws appsync graphql api resource (#617)
* feat: adding support for nuking aws appsync graphql api resource

This commit adds support for nuking AWS AppSync GraphQL APIs. The resource type is AppSyncGraphqlApi (following AWS SDK naming and casing). It supports filtering by API name and tags.

* fix: correcting casing

This commit corrects casing of structurs and functions to adhere to golang recommendations. This updates the configurable resource-type to "AppSyncGraphqlAPI".

* fix: adding apiid to properties to supporting filtering

Adding the AppSync API ID to the resource properties to support filtering by APIID.

Co-authored-by: Joshua Hornsby <josh.hornsby@availity.com>