Add tag support for IAMUser and IAMUserPolicyAtt (#762)
The tag support that was added for IAMUser wasn't working as it
was relying on the tags coming back via the UserList API. AWS
documnents that while the response includes a list of tags, it
isn't populated. Fixed this issue by calling GetUser to retrieve
the tags.

Added support for tags to IAMUserPolicyAttachments that mimics the
functionality present in IAMRolePolicyAttachments.